+++ Changelog.md
@@ -1,7 +1,7 @@
 # v3.0.1
 
 - ADDED Backup URLs to UpdateCheck (will fail less often now)
-- ADDED MP.JsonSerialize() and MP.JsonDeserialize(), which turn lua tables into json and vice-versa
+- ADDED MP.JsonEncode() and MP.JsonDecode(), which turn lua tables into json and vice-versa
 - FIXED a bug where, when run with --working-directory, the Server.log would still be in the original directory
 - FIXED a bug which could cause the plugin reload thread to spin at 100% if the reloaded plugin's didn't terminate
 - FIXED an issue which would cause servers to crash on mod download via SIGPIPE on POSIX